# swagger.yaml
openapi: 3.0.4
info:
  title: Registrar Backend API
  description: Manages the Alloy Analyzer & the model in the background.
  version: 0.1.9

servers:
  - url: "localhost:8080"

paths:
  /audits:
    get:
      summary: gets audit.
      description: Gets the audit with the specified name.
      parameters:
        - name: name
          in: query
          description: The name of the audit to retrieve.
          required: true
          schema:
            type: string
      responses:
        "200":
          description: A JSON array of audit names
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Audit'
        "400":
          description: Bad Request - The request was invalid or cannot be served.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvalidInputError'
        "404":
          description: Not Found - The specified audit does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MissingItemError'
        "500":
          description: Internal Server Error - An unexpected error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
    post:
      summary: submits a new audit.
      description: Creates a new audit with the specified name and subAudit.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Audit'
      responses:
        "201":
          description: Audit created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Audit'
        "400":
          description: Bad Request - The request was invalid or cannot be served.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvalidInputError'
        "500":
          description: Internal Server Error - An unexpected error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
  /audits/{audit}/run:
    post:
      summary: Attempts to run the audit.
      parameters:
        - name: audit
          in: path
          description: The name of the audit to run.
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunConfig'
      responses:
        "200":
          description: A JSON array of user names
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pathway'
        "400":
          description: Bad Request - The request was invalid or cannot be served.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvalidInputError'
        "500":
          description: Internal Server Error - An unexpected error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
  /courses/{code}:
    get:
      summary: gets course information.
      description: Gets the course information for the specified course code.
      parameters:
        - name: code
          in: path
          description: The course code to retrieve information for.
          required: true
          schema:
            type: string
      responses:
        "200":
          description: A JSON object containing course information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
        "400":
          description: Bad Request - The request was invalid or cannot be served.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvalidInputError'
        "404":
          description: Not Found - The specified course does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MissingItemError'
        "500":
          description: Internal Server Error - An unexpected error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'

components:
  schemas:
    SubAudit:
      type: object
      properties:
        name:
          type: string
        cardinality:
          type: string
        courses:
          type: array
          items:
            type: string
    Audit:
      type: object
      properties:
        info:
          $ref: '#/components/schemas/AuditInfo'
        subAudit:
          type: array
          items:
            $ref: '#/components/schemas/SubAudit'
    AuditInfo:
      type: object
      properties:
        gid:
          type: string
        code:
          type: string
        program:
          type: string
        isGenEd:
          type: boolean
          default: false
    SemConfig:
      type: object
      properties:
        numCourses:
          type: integer
          nullable: true
        courses:
          type: array
          minItems: 0
          maxItems: 6
          items:
            type: string
    Pathway:
      type: array
      minItems: 8
      maxItems: 8
      items:
        $ref: '#/components/schemas/SemConfig'
    Course:
      type: object
      properties:
        code:
          type: string
        dept:
          type: string
        number:
          type: integer
        name:
          type: string
        credits:
          type: float
        semester:
          type: string
        prereqs:
          type: string
    RunConfig:
      type: object
      properties:
        genEdProgram:
          type: String
        pathway:
          $ref: '#/components/schemas/Pathway'
    MissingItemError:
      type: object
      properties:
        message:
          type: string
    InvalidInputError:
      type: object
      properties:
        message:
          type: string
    InternalServerError:
      type: object
      properties:
        message:
          type: string
    